+++ src/components/chat/ChatItemOptions.vue
@@ -74,7 +74,7 @@ const applyRating = () => {
             statistics.sendEvent(new ContentEvent(
                 props.content.id,
                 props.content.name,
-                listChatKeywords(chat.messages).join(','),
+                listChatKeywords().join(','),
                 chat.id,
                 login.user?.uuid,
                 statistics.session?.id,