+++ src/views/AboutView.vue
@@ -73,7 +73,7 @@ const resetModals = () => {
         </div>
         <div class="about-container" tabindex="0">
             <div class="about-subcontainer" v-if="state.termsOpen">
-                <div class="about-content">
+                <div class="about-content" v-if="lang === 'fi'">
                     <h3>AIRe käyttöehdot</h3>
                     <p>Päivitetty 3.6.2025</p>
                     <p>
@@ -260,9 +260,171 @@ const resetModals = () => {
                         <a href="mailto:aire@jamk.fi">aire@jamk.fi</a>.
                     </p>
                 </div>
+                <div class="about-content" v-else>
+                    <h3>AIRe Platform Terms of Use</h3>
+                    <p>Updated 3 June 2025</p>
+                    <p>
+                        AIRe Platform is an AI-assisted service that serves as a simulation environment supporting
+                        Jamk’s research, development, innovation, and educational activities. AIRe helps the user
+                        identify rehabilitation needs and guides the user toward personalized rehabilitation services
+                        and content.
+                    </p>
+                    <p>
+                        By using the AIRe service, you agree to comply with these end-user terms. Use of the services is
+                        not permitted if you do not accept these terms.
+                    </p>
+                    <h4>1 General</h4>
+                    <p>
+                        These terms of use (“Terms”) apply to the AIRe online service (“Service”) provided by Jamk
+                        University of Applied Sciences Ltd, Business ID 1006550-2 (“Jamk”). AIRe allows registered
+                        users to utilize the Service to achieve positive changes in their functioning and well-being. The
+                        Service enables users (“User”) to store personal information and interact with the AI-assisted
+                        AIRe service. Users can also view rehabilitation content stored in the Service by Jamk, its
+                        partners, and other third parties. By using the Service, you consent to the use of the information
+                        you provide.
+                    </p>
+                    <p>
+                        By using AIRe, you agree to these Terms of Use and the Privacy Statement. You must accept
+                        these before using AIRe. Jamk reserves the right to amend these Terms by notifying the User.
+                        Continued use of the Service after such notification constitutes acceptance of the updated
+                        Terms.
+                    </p>
+                    <h4>2 Service provider, service language, and customer support</h4>
+                    <p>
+                        JAMK University of Applied Sciences Ltd<br />
+                        Business ID: 1006550-2<br />
+                        P.O. Box 207<br />
+                        40101 Jyväskylä, Finland
+                    </p>
+                    <p>
+                        Questions regarding the Terms of Use: 
+                        <a href="mailto:tietosuoja@jamk.fi">tietosuoja@jamk.fi</a>. The Service is available in the 
+                        language selected by the User from the options currently supported (currently Finnish, 
+                        English, Spanish, Swahili, Vietnamese, Indonesian, and Kinyarwanda).
+                    </p>
+                    <p>
+                        Customer service is available in Finnish and English, both orally and in writing.
+                    </p>
+                    <h4>3 Access to the service</h4>
+                    <p>
+                        Upon accepting these Terms of Use, Jamk grants the User a limited right to use the Service for
+                        personal, non-commercial purposes. Only users with jamk.fi credentials may register. Users
+                        must be at least 18 years old and have a valid email address.
+                    </p>
+                    <p>
+                        The Service may also be accessed using test account credentials created by an admin-level
+                        registered User. The test account allows login during a predefined period. Users logging in with a
+                        test account are bound by the same privacy and usage terms as regular Users.
+                    </p>
+                    <p>
+                        Users are responsible for arranging devices and needed connections to access the AIRe service.
+                        Devices and connections (e.g., mobile or internet) required to use the Service are not provided
+                        by Jamk. Users are responsible for acquiring, maintaining, and updating such devices and
+                        connections and for any associated costs. Users are also responsible for the security of their
+                        devices, software, and connections.
+                    </p>
+                    <h4>4 User registration</h4>
+                    <p>
+                        Use of the Service requires login as a registered user (“Create account”). Jamk allows login
+                        using a jamk.fi email address. Users create a username and password and are responsible for
+                        keeping them secure. If a User suspects their credentials have been compromised, they must
+                        change their password immediately.
+                    </p>
+                    <p>
+                        Passwords must be strong: at least 8 characters, including uppercase and lowercase letters and
+                        numbers.
+                    </p>
+                    <p>
+                        Users must always log out when using shared devices.
+                    </p>
+                    <h4>5 Pricing</h4>
+                    <p>
+                        The Service is free of charge for Users.
+                    </p>
+                    <h4>6 Data entered into the service</h4>
+                    <p>
+                        Some parts of the Service allow Users to enter data. Users are responsible for ensuring the
+                        accuracy and appropriateness of the data they provide.
+                    </p>
+                    <p>
+                        The Service may contain links to third-party websites and services. Jamk is not responsible for
+                        the content, links, products, or services of third parties.
+                    </p>
+
+                    <h4>7 Privacy and personal data</h4>
+                    <p>
+                        Jamk collects personal data and usage information from Users into the AIRe user register. Jamk
+                        processes this data in accordance with the Privacy Statement. Users agree to read and comply
+                        with the Privacy Statement.
+                    </p>
+                    <p>
+                        The processing of data, protection of privacy and personal data is governed by the AIRe Privacy
+                        Statement. By accepting these Terms of Use, the User confirms they have read and accepted
+                        the Privacy Statement.
+                    </p>
+                    <h4>8 Intellectual property rights</h4>
+                    <p>
+                        All intellectual property rights to the Service and its content belong to Jamk or third parties.
+                    </p>
+                    <p>
+                        Users do not acquire any intellectual property rights through use of the Service. Test account
+                        access is limited to personal use and test users may not make any part of the Service publicly
+                        available without written permission from Jamk.
+                    </p>
+                    <p>
+                        Jamk users may use AIRe for educational, development, and research purposes and present content 
+                        to the public.
+                    </p>
+                    <p>
+                        Users are liable for any damages caused by violating these Terms of Use, laws, regulations, or
+                        good practice. Copyright liability is governed by chapter 7 of the Copyright Act and criminal
+                        liability by chapter 49 of the Criminal Code of Finland.
+                    </p>
+                    <h4>9 Service provision</h4>
+                    <p>
+                        The Service is generally available 24/7. Jamk reserves the right to provide the Service in the
+                        manner it deems most appropriate and may change it at any time without notice. Jamk will
+                        attempt to notify Users of significant changes in advance via the Service homepage.
+                    </p>
+                    <p>
+                        Jamk may suspend the Service partially, entirely, or for specific Users. Jamk aims to minimize
+                        downtime and provide advance notice where possible.
+                    </p>
+                    <p>
+                        Jamk may suspend, cancel, terminate, or modify the AIRe service without notice. Jamk is not
+                        liable if the Service is unavailable.
+                    </p>
+                    <h4>10 Jamk’s responsibility</h4>
+                    <p>
+                        Technical maintenance is provided by GoodLife Technology on behalf of Jamk. Some content is
+                        produced by Jamk, while other content and services are provided by partners and third parties,
+                        who are responsible for their own content.
+                    </p>
+                    <p>
+                        Jamk does not guarantee uninterrupted or error-free operation. It is not responsible for the
+                        accuracy, completeness, or reliability of the content.
+                    </p>
+                    <p>
+                        Jamk is not liable for direct or indirect damages, data loss, or service interruptions.
+                    </p>
+                    <h4>11 Governing law and dispute resolution</h4>
+                    <p>
+                        These Terms of Use and the Privacy Statement are governed by Finnish law. Disputes will be
+                        resolved primarily through negotiation. If no agreement is reached, disputes will be settled in
+                        the District Court of Central Finland, except for copyright matters, which are handled by the
+                        Market Court.
+                    </p>
+                    <h4>12 Contact</h4>
+                    <p>
+                        If you have questions about these Terms of Use, contact Jamk’s data protection officer
+                        at
+                        <a href="mailto:tietosuoja@jamk.fi">tietosuoja@jamk.fi</a> or
+                        <a href="mailto:aire@jamk.fi">aire@jamk.fi</a>.
+                    </p>
+                </div>
             </div>
             <div class="about-subcontainer" v-else-if="state.privacyOpen">
-                <div class="about-content">
+                <div class="about-content" v-if="lang === 'fi'">
                     <h4>Tietosuojaseloste</h4>
                     <p>Päivitetty 3.6.2025</p>
                     <p>
@@ -483,6 +645,183 @@ const resetModals = () => {
                         <a href="https://www.jamk.fi/fi/jamk/tietosuoja-jamkissa">
                             Jamkin tietosuojalausekkeessa ja Rekisteröidyn oikeudet
                         </a>-sivulla.
+                    </p>
+                </div>
+                <div class="about-content" v-else>
+                    <h4>Privacy Statement</h4>
+                    <p>Updated 3 June 2025</p>
+                    <p>
+                        <a
+                            href="http://eur-lex.europa.eu/legal-content/FI/TXT/?uri=uriserv:OJ.L_.2016.119.01.0001.01.FIN&toc=OJ:L:2016:119:FULL">
+                            <b>EU General Data Protection Regulation (2016/679)</b>
+                        </a>
+                    </p>
+                    <h4>1 Name of the register</h4>
+                    <p>AIRe Platform</p>
+                    <h4>2 Data controller</h4>
+                    <p>
+                        Jyväskylä University of Applied Sciences (Jamk)<br />
+                        P.O. Box 207, 40101 Jyväskylä, Finland<br />
+                        +358 20 743 8100
+                    </p>
+                    <h4>3 Contact person for register matters (during office hours)</h4>
+                    <p>
+                        Responsible director of the register: Director, Institute of Rehabilitation
+                    </p>
+                    <p>
+                        Process owner of the register: Principal Specialist, Institute of Rehabilitation
+                    </p>
+                    <p>
+                        Key user: Specialist, Institute of Rehabilitation
+                    </p>
+                    <p>
+                        Technical administrator: Goodlife Technology Oy (administrator, Azure account rights), 
+                        Jamk ICT Services
+                    </p>
+                    <p>
+                        email: <a href="mailto:aire@jamk.fi">aire@jamk.fi</a><br />
+                        Jamk data protection officer: <a href="mailto:tietosuoja@jamk.fi">tietosuoja@jamk.fi</a>
+                    </p>
+                    <h4>4 Purpose of processing personal data</h4>
+                    <p>
+                        The purpose of processing personal data is to implement digital self-managed
+                        rehabilitation and its RDI and educational activities. Processing is necessary for
+                        research purposes in the public interest.
+                    </p>
+                    <p>
+                        The AIRe Platform serves as a simulation and testing environment for AI-assisted
+                        research, development, innovation, and education activities at Jamk’s Institute of
+                        Rehabilitation.
+                    </p>
+                    <p>
+                        The platform enables efficient RDI and educational operations, participant registration,
+                        smooth interaction with the AI, and reminder functionalities. It uses data provided or
+                        generated by the user to offer personalized content recommendations. Users
+                        voluntarily share information about their functioning (e.g., how a limitation affects daily
+                        activities) via AI-assisted conversation. The platform helps users identify rehabilitation
+                        needs and goals and suggests suitable content. The RDI and educational operations on
+                        the platform may include evaluating the relevance of AI-suggested themes or content
+                        and the achievement of user-defined rehabilitation goals.
+                    </p>
+                    <p>
+                        Jamk’s Institute of Rehabilitation acts as the data controller.
+                    </p>
+                    <p>
+                        This privacy statement may be supplemented with specific privacy statements for
+                        particular processing purposes.
+                    </p>
+                    <h4>5 Data content of the register</h4>
+                    <p>
+                        Registered users include Jamk staff and students, as well as external volunteers
+                        participating in RDI or educational activities. Research participants will complete a
+                        separate consent form. Each project using the platform is responsible for its own
+                        privacy notice.
+                    </p>
+                    <p>
+                        Basic data the platform collects includes:
+                    </p>
+                    <ol>
+                        <li>
+                            First name, last name, gender, year of birth, country, and email address (email
+                            stored only for Jamk users and technical admins; external users log in with
+                            anonymized credentials).
+                        </li>
+                        <li>
+                            Voluntarily shared data with the AI- assisted conversation (e.g., descriptions of
+                            limitations in functioning).
+                        </li>
+                        <li>
+                            Preferences, restrictions, consents, and other user choices (currently not in use).
+                        </li>
+                        <li>
+                            Completed surveys, uploaded files, logs, and message timestamps.
+                        </li>
+                        <li>
+                            Technical identifiers: IP address, API path, device and OS identifiers, random unique ID, 
+                            hashed email, login times, role and access rights, email confirmation status, username, 
+                            visited pages, browser type, referring URL, and server source.
+                        </li>
+                    </ol>
+                    <h4>6 Legal basis for processing personal data</h4>
+                    <ul>
+                        <li>Consent of the data subject</li>
+                        <li>Public interest and official authority</li>
+                        <li>EU GDPR (2016/679)</li>
+                        <li>Finnish Data Protection Act (1050/2018)</li>
+                        <li>Act on the Openness of Government Activities (621/1999)</li>
+                    </ul>
+                    <h4>7 Regular sources of data</h4>
+                    <p>
+                        Data is obtained directly from the data subject.
+                    </p>
+                    <h4>8 Regular data disclosures and data transfers outside the EU/EEA</h4>
+                    <p>
+                        AIRe uses Microsoft Azure cloud services, all hosted within the EU. The Azure OpenAI 
+                        interface does not use conversations for training or share user input with third parties.
+
+                        <a
+                            href="https://learn.microsoft.com/en-us/legal/cognitive-services/openai/data-privacy?tabs=azure-portal">
+                            Data, privacy, and security for Azure OpenAI Service
+                        </a>
+                    </p>
+                    <p>
+                        Data is not regularly disclosed outside Jamk/AIRe. Transfers outside the EU/EEA may occur in 
+                        international research projects, based on specific agreements, consent, or legal grounds, 
+                        and are defined in project-specific privacy notices.
+                    </p>
+                    <h4>9 Principles of data protection</h4>
+                    <ul>
+                        <li>
+                            Data is stored securely to prevent unauthorized access, loss, alteration, or
+                            unlawful processing.
+                        </li>
+                        <li>
+                            Admin users can only access data necessary for their tasks.
+                        </li>
+                        <li>
+                            Documents containing personal data are properly deleted.
+                        </li>
+                        <li>
+                            Network and server equipment are in locked Jamk facilities with restricted 
+                            access.
+                        </li>
+                        <li>
+                            Access rights are role-based and limited to authorized personnel.
+                        </li>
+                        <li>
+                            Jamk staff are bound by confidentiality agreements.
+                        </li>
+                    </ul>
+                    <h4>10 Data retention</h4>
+                    <p>
+                        Data is retained as long as the user has an account or as required by research.
+                    </p>
+                    <p>
+                        Admins can deactivate accounts but not delete them.
+                    </p>
+                    <p>
+                        Users can delete their accounts and choose whether to leave anonymized data in 
+                        the system.
+                    </p>
+                    <p>
+                        Logs (anonymized) retain IP, API path, timestamps, hashed email, and access
+                        metadata.
+                    </p>
+                    <p>
+                        All other data is encrypted and visible only to the user or someone with their
+                        credentials.
+                    </p>
+                    <h4>11 Profiling and automated decision-making</h4>
+                    <p>
+                        No automated decisions or profiling are performed. If profiling is used in research, it will
+                        be disclosed in a separate privacy notice.
+                    </p>
+                    <h4>12 Data subject rights</h4>
+                    <p>
+                        More information on data protection and data subject rights is available in
+                        <a href="https://www.jamk.fi/en/data-protection-in-jamk">
+                            Jamk’s Data Protection and the Data Subject Rights
+                        </a> page.
                     </p>
                 </div>
             </div>